requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
langchain-pinecone
langchain_community
langchain_openai
-e .